requirements.txt
streamlit>=1.31.0
python-dotenv>=1.0.0
PyPDF2>=3.0.1
langchain>=0.1.0
openai>=1.0.0
faiss-cpu>=1.7.4
gpt4all>=2.2.0